LayoutTests/imported/w3c/web-platform-tests/webrtc/RTCRtpTransceiver.https-expected.txt


PASS checkAddTransceiverNoTrack
PASS checkAddTransceiverWithTrack
PASS checkAddTransceiverWithAddTrack
PASS checkAddTransceiverWithDirection
FAIL checkMsidNoTrackId promise_test: Unhandled rejection with value: object "SyntaxError: Expects 2 fields."
PASS checkAddTransceiverWithSetRemoteOfferSending
PASS checkAddTransceiverWithSetRemoteOfferNoSend
PASS checkAddTransceiverBadKind
PASS checkNoMidOffer
PASS checkNoMidAnswer
PASS checkSetDirection
FAIL checkCurrentDirection assert_equals: expected "[{currentDirection:null}]" but got "[{currentDirection:\"inactive\"}]"
PASS checkSendrecvWithNoSendTrack
PASS checkSendrecvWithTracklessStream
PASS checkAddTransceiverNoTrackDoesntPair
PASS checkAddTransceiverWithTrackDoesntPair
PASS checkAddTransceiverThenReplaceTrackDoesntPair
FAIL checkAddTransceiverThenAddTrackPairs assert_equals: expected "[{sender:{track:{}}}]" but got "[{sender:{track:{}}},{}]"
PASS checkAddTrackPairs
PASS checkReplaceTrackNullDoesntPreventPairing
PASS checkRemoveAndReadd
PASS checkAddTrackExistingTransceiverThenRemove
FAIL checkRemoveTrackNegotiation assert_equals: Rollback should have added two tracks expected 2 but got 0
FAIL checkMute assert_true: expected true got false
FAIL checkStop assert_equals: expected "[{currentDirection:\"sendrecv\",direction:\"stopped\",receiver:{track:{kind:\"audio\",readyState:\"ended\"}},sender:{track:{kind:\"audio\"}}}]" but got "[{currentDirection:\"sendrecv\",direction:\"inactive\",receiver:{track:{kind:\"audio\",readyState:\"ended\"}},sender:{track:null}}]"
FAIL checkStopAfterCreateOffer assert_equals: expected "[{currentDirection:\"sendrecv\",direction:\"stopped\"}]" but got "[{currentDirection:\"sendrecv\",direction:\"inactive\"}]"
FAIL checkStopAfterSetLocalOffer assert_equals: expected "[{currentDirection:\"sendrecv\",direction:\"stopped\"}]" but got "[{currentDirection:\"sendrecv\",direction:\"inactive\"}]"
FAIL checkStopAfterSetRemoteOffer assert_equals: expected "stopped" but got "inactive"
FAIL checkStopAfterCreateAnswer assert_equals: expected "[{currentDirection:null,direction:\"stopped\"}]" but got "[{currentDirection:null,direction:\"inactive\"}]"
FAIL checkStopAfterSetLocalAnswer assert_equals: expected "[{currentDirection:\"sendrecv\",direction:\"stopped\"}]" but got "[{currentDirection:\"sendrecv\",direction:\"inactive\"}]"
PASS checkStopAfterClose
FAIL checkLocalRollback assert_equals: expected "[{direction:\"stopped\"}]" but got "[{direction:\"inactive\"}]"
PASS checkRollbackAndSetRemoteOfferWithDifferentType
FAIL checkRemoteRollback assert_equals: expected "[]" but got "[{}]"
FAIL checkMsectionReuse assert_equals: expected "[{currentDirection:null,mid:null}]" but got "[{currentDirection:\"inactive\",mid:\"0\"}]"
PASS checkStopAfterCreateOfferWithReusedMsection
PASS checkAddIceCandidateToStoppedTransceiver
PASS checkBundleTagRejected
